backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
boto3>=1.34.0
botocore>=1.34.0
python-dotenv>=1.0.0
strands-agents>=1.0.0
strands-agents-tools>=0.2.0
